move snapshots to its own file
